add CADO_DEBUG for bwc-ptrace.sh

+++ tests/linalg/bwc/bwc-tests.sh
@@ -11,6 +11,8 @@ matdir=`mktemp -d /tmp/bwc-test-mats.XXXXXXXX`
 
 $script mats=$matdir bindir=$bindir wipe=1 wdir=$wdir nomagma=1 "$@"
 
-rm -rf $wdir
-rm -rf $matdir
+if ! [ "$CADO_DEBUG" ] ; then
+    rm -rf $wdir
+    rm -rf $matdir
+fi
